+++ csharp/ToolGood.FlowVision/ToolGood.Algorithm2/AlgorithmEngine.cs
@@ -607,7 +607,7 @@ namespace ToolGood.Algorithm2
 						LastError = obj.ErrorMsg;
 						return def;
 					}
-					return obj.NumberValue;
+					return (double)obj.NumberValue;
 				}
 			} catch (Exception ex) {
 				LastError = ex.Message + "\r\n" + ex.StackTrace;
